Switch to CMake
There's really no reason for this project not to support more build
environments on more platforms in the future.

The two .mak files were originally written for RSXNT, an initial effort of
porting GCC to Win32, which was discontinued in 1999 and then got superseded
by MinGW in the early 2000s. These files don't work with any other make tool I
have, I couldn't get them to work with the final release of RSXNT, and the
only advantage of RSXNT (that is, console executables working under DOS as
well) does not matter at all in this day and age. (Even if your applications
need to target DOS, cross-compiling is by far the most convenient option.)

On the other hand, the original Makefile built fine with current versions of
GCC on both Linux and MinGW. Continuing to maintain it would have been an
option, but it still had the obvious flaw of not being able to detect the few
compiler-dependent CRT functions needed by the code.
Keeping it would now have required a platform-specific alink_config.h in the
root of the source tree, which might have clashed with the custom version for
the different CMake configurations. Therefore, it seems better to just
completely ditch all old Makefiles in favor of consistent, out-of-tree CMake
builds.

As for the tests in the original Makefile... well, some of the required .obj
files seem have vanished from the Internet. Better to write new ones, then.